Move TTL to testcase/
