+++ run/workflow/ttl.jsonnet
@@ -1,0 +1,11 @@
+{
+  _config+:: {
+    case_name: 'ttl',
+    image_name: 'hub.pingcap.net/qa/tipocket',
+    args+: {
+      // k8s configurations
+      // 'storage-class': 'local-storage',
+    },
+    command: {},
+  },
+}
